+++ utils/validate_test.go
@@ -4,20 +4,21 @@ import "testing"
 
 func TestStrIsInt(t *testing.T) {
 	tc := []struct {
+		name     string
 		inp      string
 		expected bool
 	}{
-		{"anna", false},
-		{"123", true},
-		{"-123", true},
-		{"0", true},
-		{"-1", true},
-		{"123.23", false},
-		{"123.0", false},
+		{"Non-integer string", "anna", false},
+		{"Positive integer", "123", true},
+		{"Negative integer", "-123", true},
+		{"Zero as integer", "0", true},
+		{"Negative one", "-1", true},
+		{"Decimal number with fraction", "123.23", false},
+		{"Decimal number ending with zero", "123.0", false},
 	}
 
 	for _, tt := range tc {
-		t.Run("Test case: "+tt.inp, func(t *testing.T) {
+		t.Run(tt.name, func(t *testing.T) {
 			got := StrIsInt(tt.inp)
 
 			if got != tt.expected {