+++ evaluator/func.go
@@ -18,6 +18,7 @@ var functions = map[object.ObjectType]map[string]*object.Builtin{
 		"truncate":   {Fn: strTruncateFunc},
 		"decimal":    {Fn: strDecimalFunc},
 		"at":         {Fn: strAtFunc},
+		"first":      {Fn: strFirstFunc},
 	},
 	object.ARR_OBJ: {
 		"len":     {Fn: arrayLenFunc},