+++ evaluator/array_func.go
@@ -152,3 +152,27 @@ func arrayShuffleFunc(_ *ctx.EvalCtx, receiver object.Object, _ ...object.Object
 
 	return &object.Array{Elements: shuffled}, nil
 }
+
+// arrayContainsFunc checks if the given array contains the given element
+func arrayContainsFunc(_ *ctx.EvalCtx, receiver object.Object, args ...object.Object) (object.Object, error) {
+	if len(args) == 0 {
+		msg := fmt.Sprintf(fail.ErrFuncRequiresOneArg, "contains", object.ARR_OBJ)
+		return nil, errors.New(msg)
+	}
+
+	elems := receiver.(*object.Array).Elements
+
+	if len(elems) == 0 {
+		return &object.Bool{Value: false}, nil
+	}
+
+	targetVal := args[0].Val()
+
+	for _, el := range elems {
+		if el.Val() == targetVal {
+			return &object.Bool{Value: true}, nil
+		}
+	}
+
+	return &object.Bool{Value: false}, nil
+}