+++ evaluator/array_func_test.go
@@ -1,7 +1,6 @@
 package evaluator
 
 import (
-	"strings"
 	"testing"
 )
 
@@ -44,44 +43,14 @@ func TestEvalArrayFunctions(t *testing.T) {
 		{`{{ [].shuffle() }}`, ""},
 		{`{{ [1].shuffle() }}`, "1"},
 		{`{{ ["test"].shuffle() }}`, "test"},
+		// contains
+		{`{{ [].contains(1) }}`, "0"},
+		{`{{ [1, 2, 3].contains(1) }}`, "1"},
+		{`{{ [1, 2, 3].contains(4) }}`, "0"},
+		{`{{ [{}, [1, 2], 3].contains(1) }}`, "0"},
 	}
 
 	for _, tc := range tests {
 		evaluationExpected(t, tc.inp, tc.expected)
-	}
-}
-
-func TestArrayShuffle(t *testing.T) {
-	input := `{{ [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15].shuffle() }}`
-	initial := "0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15"
-
-	shuffled := testEval(input).String()
-
-	// Ensure the shuffled array has the same length as the original
-	if len(shuffled) != len(initial) {
-		t.Errorf("expected shuffled array to have length %d, got %d", len(initial), len(shuffled))
-	}
-
-	// Check that the shuffle changed the order
-	if shuffled == initial {
-		t.Error("expected shuffled array to be different from the initial array")
-	}
-
-	// Check that all numbers from the initial array are present in the shuffled array
-	initialElements := strings.Split(initial, ", ")
-	shuffledElements := strings.Split(shuffled, ", ")
-
-	// Create a map to track occurrences of elements in the shuffled array
-	elementMap := make(map[string]bool)
-
-	for _, elem := range shuffledElements {
-		elementMap[elem] = true
-	}
-
-	// Ensure all elements from the initial array are in the shuffled array
-	for _, elem := range initialElements {
-		if !elementMap[elem] {
-			t.Errorf("expected element %s to be present in shuffled array", elem)
-		}
 	}
 }